+++ test/reference/fill.hpp
@@ -22,7 +22,7 @@ class DataFormat;
 /// @param[in] seed Random seed.
 ///
 /// @return The data buffer for the matrix.
-std::vector<uint8_t> fill_matrix_random(size_t height, size_t width, const DataFormat& format, uint64_t seed);
+std::vector<uint8_t> fill_matrix_random(size_t height, size_t width, const DataFormat& format, uint32_t seed);
 
 /// Creates a new data buffer filled with random data.
 ///
@@ -33,6 +33,6 @@ std::vector<uint8_t> fill_matrix_random(size_t height, size_t width, const DataF
 ///
 /// @return The data buffer.
 template <typename Value>
-std::vector<uint8_t> fill_random(size_t length, uint64_t seed);
+std::vector<uint8_t> fill_random(size_t length, uint32_t seed);
 
 }  // namespace kai::test
